Merge IoT and AI to take smart decisions.
We are in the era of Industry 4.0, and with that, comes the threat of increasing global competition. To be more competitive, it is crucial that organizations especially in the manufacturing industry can apply the tools of IoT and AI, which are at the forefront of aiding organizations to meet their long-term goals. IoT or Internet of Things is the name given to the network of physical devices that are connected to the internet. This network of connected objects includes cloud-connected devices, sensors, and industrial equipment. IoT or the Internet of things is a network of technologies or sensors that consists of some advanced technologies, that are embedded into it. It helps in interacting and communicating with their data. This process involves receiving and transferring data through the network without any human intervention. These data from the devices or the sensors can be stored in the cloud and again it can be made available for real-time analytics. Thus, IoT collects this data from different environments which is generally vast, due to the ever-rapid growing of devices and sensors. These big data are important only when it is transformed into a piece of valuable and actionable information within a specified time period. Here comes the role of artificial intelligence or AI. It collects the data by applying analytics and it converts these collective data into applications to extract the meaning from it. When this big data is fed from IoT devices and into an AI system, it reviews and analyzes the data, and therefore it produces decisions made either by machines or humans. The importance of big data is significant only when it's transformed into relative, valuable, and actionable information within a specified time period. Artificial Intelligence is used for learning and thus making self-decisions by using the process of complex organized data and also unorganized data. The IoT and AI play a crucial role in the future as it has a growing need for technologies in both government and private sectors, especially in the manufacturing industry. At this point, edge analytics is presented by Brabo Edge Platform, a manufacturing connectivity and intelligence platform developed by Solulever, a Dutch technology startup.